Refactor placement of tests into folder
