+++ tests/integration/test_compile.sh
